+++ tests/Input/Button/AttributeTest.php
@@ -133,6 +133,16 @@ final class AttributeTest extends TestCase
         );
     }
 
+    public function testGenerateId(): void
+    {
+        $this->assertStringContainsString('id="button-', Button::widget()->render());
+    }
+
+    public function testGetValue(): void
+    {
+        $this->assertSame('value', Button::widget()->value('value')->getValue());
+    }
+
     public function testHidden(): void
     {
         Assert::equalsWithoutLE(
