+++ tests/Input/Checkbox/AttributeTest.php
@@ -153,6 +153,16 @@ final class AttributeTest extends TestCase
         );
     }
 
+    public function testGenerateId(): void
+    {
+        $this->assertStringContainsString('id="checkbox-', Checkbox::widget()->render());
+    }
+
+    public function testGetValue(): void
+    {
+        $this->assertSame('value', Checkbox::widget()->value('value')->getValue());
+    }
+
     public function testHidden(): void
     {
         Assert::equalsWithoutLE(
