+++ tests/Input/Datetime/AttributeTest.php
@@ -113,6 +113,16 @@ final class AttributeTest extends TestCase
         );
     }
 
+    public function testGenerateId(): void
+    {
+        $this->assertStringContainsString('id="datetime-', Datetime::widget()->render());
+    }
+
+    public function testGetValue(): void
+    {
+        $this->assertSame('value', Datetime::widget()->value('value')->getValue());
+    }
+
     public function testHidden(): void
     {
         Assert::equalsWithoutLE(
