+++ tests/Input/Image/AttributeTest.php
@@ -11,7 +11,7 @@ use PHPUnit\Framework\TestCase;
 /**
  * @psalm-suppress PropertyNotSetInConstructor
  */
-final class RenderTest extends TestCase
+final class AttributeTest extends TestCase
 {
     public function testAlt(): void
     {
@@ -213,79 +213,6 @@ final class RenderTest extends TestCase
         );
     }
 
-    public function testPrefix(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            prefix
-            <input id="image-65a15e0439570" type="image">
-            HTML,
-            Image::widget()->id('image-65a15e0439570')->prefix('prefix')->render()
-        );
-    }
-
-    public function testPrefixContainer(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <div>
-            prefix
-            </div>
-            <input id="image-65a15e0439570" type="image">
-            HTML,
-            Image::widget()->id('image-65a15e0439570')->prefix('prefix')->prefixContainer(true)->render()
-        );
-    }
-
-    public function testPrefixContainerAttributes(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <div class="value">
-            prefix
-            </div>
-            <input id="image-65a15e0439570" type="image">
-            HTML,
-            Image::widget()->id('image-65a15e0439570')
-                ->prefix('prefix')
-                ->prefixContainer(true)
-                ->prefixContainerAttributes(['class' => 'value'])
-                ->render()
-        );
-    }
-
-    public function testPrefixContainerClass(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <div class="value">
-            prefix
-            </div>
-            <input id="image-65a15e0439570" type="image">
-            HTML,
-            Image::widget()->id('image-65a15e0439570')
-                ->prefix('prefix')
-                ->prefixContainer(true)
-                ->prefixContainerClass('value')
-                ->render()
-        );
-    }
-
-    public function testPrefixContainerTag(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <span>prefix</span>
-            <input id="image-65a15e0439570" type="image">
-            HTML,
-            Image::widget()->id('image-65a15e0439570')
-                ->prefix('prefix')
-                ->prefixContainer(true)
-                ->prefixContainerTag('span')
-                ->render()
-        );
-    }
-
     public function testReadonly(): void
     {
         Assert::equalsWithoutLE(
@@ -326,79 +253,6 @@ final class RenderTest extends TestCase
         );
     }
 
-    public function testSuffix(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="image-65a15e0439570" type="image">
-            suffix
-            HTML,
-            Image::widget()->id('image-65a15e0439570')->suffix('suffix')->render()
-        );
-    }
-
-    public function testSuffixContainer(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="image-65a15e0439570" type="image">
-            <div>
-            suffix
-            </div>
-            HTML,
-            Image::widget()->id('image-65a15e0439570')->suffix('suffix')->suffixContainer(true)->render()
-        );
-    }
-
-    public function testSuffixContainerAttributes(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="image-65a15e0439570" type="image">
-            <div class="value">
-            suffix
-            </div>
-            HTML,
-            Image::widget()->id('image-65a15e0439570')
-                ->suffix('suffix')
-                ->suffixContainer(true)
-                ->suffixContainerAttributes(['class' => 'value'])
-                ->render()
-        );
-    }
-
-    public function testSuffixContainerClass(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="image-65a15e0439570" type="image">
-            <div class="value">
-            suffix
-            </div>
-            HTML,
-            Image::widget()->id('image-65a15e0439570')
-                ->suffix('suffix')
-                ->suffixContainer(true)
-                ->suffixContainerClass('value')
-                ->render()
-        );
-    }
-
-    public function testSuffixContainerTag(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="image-65a15e0439570" type="image">
-            <span>suffix</span>
-            HTML,
-            Image::widget()->id('image-65a15e0439570')
-                ->suffix('suffix')
-                ->suffixContainer(true)
-                ->suffixContainerTag('span')
-                ->render()
-        );
-    }
-
     public function testTabindex(): void
     {
         Assert::equalsWithoutLE(
@@ -406,18 +260,6 @@ final class RenderTest extends TestCase
             <input id="image-65a15e0439570" type="image" tabindex="1">
             HTML,
             Image::widget()->id('image-65a15e0439570')->tabIndex(1)->render()
-        );
-    }
-
-    public function testTemplate(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <div>
-            <input id="image-65a15e0439570" type="image">
-            </div>
-            HTML,
-            Image::widget()->id('image-65a15e0439570')->template('<div>\n{tag}\n</div>')->render()
         );
     }
 
@@ -448,6 +290,26 @@ final class RenderTest extends TestCase
             <input id="image-65a15e0439570" type="image">
             HTML,
             Image::widget()->id('image-65a15e0439570')->value(null)->render()
+        );
+    }
+
+    public function testWithoutId(): void
+    {
+        Assert::equalsWithoutLE(
+            <<<HTML
+            <input type="image">
+            HTML,
+            Image::widget()->id(null)->render()
+        );
+    }
+
+    public function testWithoutName(): void
+    {
+        Assert::equalsWithoutLE(
+            <<<HTML
+            <input id="image-65a15e0439570" type="image">
+            HTML,
+            Image::widget()->id('image-65a15e0439570')->name(null)->render()
         );
     }
 }
